Convert window_digit_b2d_lookup to UINT8
Issue #119
-e 
Original/build commit ID: de847b058c4ca7d48694e4344862dcc543c635d6

+++ window.c
@@ -15,7 +15,7 @@
 #include "main.h"
 #include "menu_config.h"
 
-const unsigned short window_digit_b2d_lookup[9U][32U] = {
+const UINT8 window_digit_b2d_lookup[9U][32U] = {
   // digit 1
   {1, 2, 4, 8, 6, 2, 4, 8, 6, 2, 4, 8, 6, 2, 4, 8, 6, 2, 4, 8, 6, 2, 4, 8, 6, 2, 4, 8, 6, 2, 4, 8},
   // digit 2
